feat: unwrap skill namespace to flat exports + barrel
